# config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Student Management System API",
    "version": "1.0.0",
    "description": "Interactive Swagger UI documentation for all endpoints of the Student Management System backend."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token obtained from /api/auth/login"
      }
    }
  },
...
  "paths": {
...
    "/api/students/{id}/internships": {
      "get": {
        "tags": ["Students"],
        "summary": "Get student internships",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "student_1" } }
        ],
        "responses": {
          "200": { "description": "List of internships" }
        }
      }
    },
...
    "/api/faculty": {
      "get": {
        "tags": ["Faculty"],
        "summary": "List faculty directory",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "List of faculty" }
        }
      }
    },
    "/api/faculty/{id}": {
      "get": {
        "tags": ["Faculty"],
        "summary": "Get faculty by ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "faculty_1" } }
        ],
        "responses": {
          "200": { "description": "Faculty profile" }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Get user notifications",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "List of user alerts" }
        }
      }
    },
    "/api/notifications/read-all": {
      "put": {
        "tags": ["Notifications"],
        "summary": "Mark all notifications as read",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "All marked as read" }
        }
      }
    },
    "/api/notifications/unread": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Get unread notifications for logged-in user",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "List of unread notifications" }
        }
      }
    },
    "/api/notifications/{id}/read": {
      "put": {
        "tags": ["Notifications"],
        "summary": "Mark a single notification as read",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Notification marked as read" },
          "404": { "description": "Notification not found" }
        }
      }
    },
    "/api/notifications/{id}": {
      "delete": {
        "tags": ["Notifications"],
        "summary": "Delete a notification belonging to user",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Notification deleted successfully" },
          "404": { "description": "Notification not found" }
        }
      }
    },
...
    "/api/internships": {
      "get": {
        "tags": ["Internships & Achievements"],
        "summary": "Get internships",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "List of internships" }
        }
      },
      "post": {
        "tags": ["Internships & Achievements"],
        "summary": "Create internship (Student only)",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "201": { "description": "Internship created" }
        }
      }
    },
    "/api/internships/{id}": {
      "delete": {
        "tags": ["Internships & Achievements"],
        "summary": "Delete internship (Owner student or Admin)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Internship deleted successfully" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
    },
...
  }
}